refactor: api route modification
